refactor: rename apiApp.ts -> apiCreateApp.ts
